back-end/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.30
